moved smoothness tests under exclude_unpaved
