+++ test/gurka/test_exclude_unpaved_smoothness.cc
@@ -13,8 +13,138 @@ using namespace valhalla::baldr;
 using namespace valhalla::gurka;
 using namespace valhalla::mjolnir;
 
-const std::unordered_map<std::string, std::string> build_config{
-    {"mjolnir.admin", {VALHALLA_SOURCE_DIR "test/data/netherlands_admin.sqlite"}}};
+namespace {
+const std::vector<std::string> kSupportedCostingModels = {
+    "auto",
+    "taxi",
+    "bus",
+    "truck",
+};
+} // namespace
+
+class ExcludeUnpavedTest : public ::testing::Test {
+protected:
+  static gurka::map map;
+
+  static void SetUpTestSuite() {
+    constexpr double grid_size_meters = 100.;
+
+    const std::string ascii_map = R"(
+   E----F----G----H----I----A----J----K----L
+                                 |    |
+                                 |    |
+                                 |    |
+                                 |    |
+                                 M----N
+    )";
+
+    const gurka::ways ways = {
+        {"EF", {{"highway", "residential"}, {"surface", "compacted"}}},
+        {"FG", {{"highway", "residential"}, {"surface", "compacted"}}},
+        {"GH", {{"highway", "residential"}}},
+        {"HI", {{"highway", "residential"}, {"surface", "unpaved"}}},
+        {"IA", {{"highway", "residential"}, {"surface", "unpaved"}}},
+        {"IJ", {{"highway", "residential"}}},
+        {"JK", {{"highway", "residential"}, {"surface", "gravel"}}},
+        {"KL", {{"highway", "residential"}}},
+        {"JM", {{"highway", "residential"}}},
+        {"MN", {{"highway", "residential"}}},
+        {"NK", {{"highway", "residential"}}},
+    };
+
+    const auto layout = gurka::detail::map_to_coordinates(ascii_map, grid_size_meters);
+    map = gurka::buildtiles(layout, ways, {}, {}, "test/data/exclude_unpaved");
+  }
+};
+
+gurka::map ExcludeUnpavedTest::map = {};
+
+TEST_F(ExcludeUnpavedTest, UnpavedRoadsInTheMiddle) {
+  // Without options
+  for (const auto& costing : kSupportedCostingModels) {
+    const auto result = gurka::do_action(valhalla::Options::route, map, {"I", "L"}, costing);
+    gurka::assert::raw::expect_path(result, {"IJ", "JK", "KL"});
+  }
+
+  // Use unpaved roads
+  for (const auto& costing : kSupportedCostingModels) {
+    const auto result = gurka::do_action(valhalla::Options::route, map, {"I", "L"}, costing,
+                                         {{"/costing_options/" + costing + "/exclude_unpaved", "0"}});
+    gurka::assert::raw::expect_path(result, {"IJ", "JK", "KL"});
+  }
+
+  // Do not use unpaved roads
+  for (const auto& costing : kSupportedCostingModels) {
+    const auto result = gurka::do_action(valhalla::Options::route, map, {"I", "L"}, costing,
+                                         {{"/costing_options/" + costing + "/exclude_unpaved", "1"}});
+    gurka::assert::raw::expect_path(result, {"IJ", "JM", "MN", "NK", "KL"});
+  }
+}
+
+TEST_F(ExcludeUnpavedTest, UnpavedRoadsUnsupported) {
+  const std::string start = "E";
+  const std::string end = "L";
+  for (const auto& costing : std::vector<std::string>{"bicycle", "pedestrian"}) {
+    const auto result_0 =
+        gurka::do_action(valhalla::Options::route, map, {start, end}, costing,
+                         {{"/costing_options/" + costing + "/exclude_unpaved", "1"}});
+    EXPECT_EQ(result_0.trip().routes_size(), 1);
+    const auto result_1 =
+        gurka::do_action(valhalla::Options::route, map, {start, end}, costing,
+                         {{"/costing_options/" + costing + "/exclude_unpaved", "0"}});
+    EXPECT_EQ(result_1.trip().routes_size(), 1);
+    EXPECT_EQ(gurka::detail::get_paths(result_0), gurka::detail::get_paths(result_1));
+  }
+
+  // motor_scooter, motorcycle are unsupported costing models too, but the engine does not get routes
+  // through unpaved roads. These edges are not allowed in the Allowed and ReverseAllowed methods. It
+  // includes only roads that have surface greater than Surface::kDirt(See kMinimumMotorcycleSurface,
+  // kMinimumScooterSurface constants).
+}
+
+TEST_F(ExcludeUnpavedTest, UnpavedRoadsInTheBeginning) {
+  // Without options
+  for (const auto& costing : kSupportedCostingModels) {
+    const auto result = gurka::do_action(valhalla::Options::route, map, {"E", "H"}, costing);
+    gurka::assert::raw::expect_path(result, {"EF", "FG", "GH"});
+  }
+
+  // Use unpaved roads
+  for (const auto& costing : kSupportedCostingModels) {
+    const auto result = gurka::do_action(valhalla::Options::route, map, {"E", "H"}, costing,
+                                         {{"/costing_options/" + costing + "/exclude_unpaved", "0"}});
+    gurka::assert::raw::expect_path(result, {"EF", "FG", "GH"});
+  }
+
+  // Do not use unpaved roads
+  for (const auto& costing : kSupportedCostingModels) {
+    const auto result = gurka::do_action(valhalla::Options::route, map, {"E", "H"}, costing,
+                                         {{"/costing_options/" + costing + "/exclude_unpaved", "1"}});
+    gurka::assert::raw::expect_path(result, {"EF", "FG", "GH"});
+  }
+}
+
+TEST_F(ExcludeUnpavedTest, UnpavedRoadsInTheEnd) {
+  // Without options
+  for (const auto& costing : kSupportedCostingModels) {
+    const auto result = gurka::do_action(valhalla::Options::route, map, {"G", "A"}, costing);
+    gurka::assert::raw::expect_path(result, {"GH", "HI", "IA"});
+  }
+
+  // Use unpaved roads
+  for (const auto& costing : kSupportedCostingModels) {
+    const auto result = gurka::do_action(valhalla::Options::route, map, {"G", "A"}, costing,
+                                         {{"/costing_options/" + costing + "/exclude_unpaved", "0"}});
+    gurka::assert::raw::expect_path(result, {"GH", "HI", "IA"});
+  }
+
+  // Do not use unpaved roads
+  for (const auto& costing : kSupportedCostingModels) {
+    const auto result = gurka::do_action(valhalla::Options::route, map, {"G", "A"}, costing,
+                                         {{"/costing_options/" + costing + "/exclude_unpaved", "1"}});
+    gurka::assert::raw::expect_path(result, {"GH", "HI", "IA"});
+  }
+}
 
 valhalla::gurka::map BuildPBF(const std::string& workdir) {
   const std::string ascii_map = R"(
@@ -124,8 +254,7 @@ TEST(Standalone, SmoothnessWithSurface) {
     constexpr double gridsize = 100;
     const auto layout =
         gurka::detail::map_to_coordinates(ascii_map, gridsize, {5.1079374, 52.0887174});
-    auto map = gurka::buildtiles(layout, ways, {}, {}, "test/data/gurka_smoothness_with_surface",
-                                 build_config);
+    auto map = gurka::buildtiles(layout, ways, {}, {}, "test/data/gurka_smoothness_with_surface");
 
     GraphReader graph_reader(map.config.get_child("mjolnir"));
 
@@ -175,8 +304,7 @@ TEST(Standalone, SmoothnessNoSurface) {
     constexpr double gridsize = 100;
     const auto layout =
         gurka::detail::map_to_coordinates(ascii_map, gridsize, {5.1079374, 52.0887174});
-    auto map = gurka::buildtiles(layout, ways, {}, {}, "test/data/gurka_smoothness_no_surface",
-                                 build_config);
+    auto map = gurka::buildtiles(layout, ways, {}, {}, "test/data/gurka_smoothness_no_surface");
 
     GraphReader graph_reader(map.config.get_child("mjolnir"));
 
@@ -245,8 +373,7 @@ TEST(Standalone, SmoothnessWithTrackType) {
     constexpr double gridsize = 100;
     const auto layout =
         gurka::detail::map_to_coordinates(ascii_map, gridsize, {5.1079374, 52.0887174});
-    auto map = gurka::buildtiles(layout, ways, {}, {}, "test/data/gurka_smoothness_no_surface",
-                                 build_config);
+    auto map = gurka::buildtiles(layout, ways, {}, {}, "test/data/gurka_smoothness_no_surface");
 
     GraphReader graph_reader(map.config.get_child("mjolnir"));
